move items transform steps into extra file
