+++ dlt/extract/validation.py
@@ -8,7 +8,8 @@ except ModuleNotFoundError:
 
 from dlt.common.typing import TDataItems
 from dlt.common.schema.typing import TAnySchemaColumns, TSchemaContract, TSchemaEvolutionMode
-from dlt.extract.items import TTableHintTemplate, ValidateItem
+from dlt.extract.items import TTableHintTemplate
+from dlt.extract.items_transform import ValidateItem
 
 
 _TPydanticModel = TypeVar("_TPydanticModel", bound=PydanticBaseModel)
